add PH embed to LP
